refactor(frontend): update app structure and routing
Move views to a folder, add React Router, and adjust Order view to new routes.
