+++ src/main.tsx
@@ -2,7 +2,7 @@ import {StrictMode} from 'react';
 import {createRoot} from 'react-dom/client';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.scss';
-import App from './App';
+import App from "./App.tsx";
 
 createRoot(document.getElementById('root')!).render(
     <StrictMode>
