--- src/util/constants.ts
@@ -1,19 +1,0 @@
-export const ORDER_STATES = {
-    PENDING: 'Pendiente',
-    READY: 'Listo',
-    DELIVERED: 'Entregado',
-    CANCELLED: 'Cancelado'
-};
-
-export const API_ENDPOINTS = {
-    ORDERS: '/pedidos',
-    SEARCH_ORDERS: '/pedidos/search',
-    UPDATE_ORDER_STATUS: '/pedidos'
-};
-
-export const TOAST_TYPES = {
-    SUCCESS: 'success',
-    ERROR: 'danger',
-    WARNING: 'warning',
-    INFO: 'info'
-};
